appeals_novoross/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-filter==2.3.0
django-mailbox==4.8.1
django-model-utils==4.0.0
django-phone-field==1.8.1
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
idna==2.10
Markdown==3.2.2
psycopg2-binary==2.8.5
pytz==2020.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
